+++ src/test/java/shop/mtcoding/miniproject2/controller/CompanyResumeControllerTest.java
@@ -1,8 +1,7 @@
 package shop.mtcoding.miniproject2.controller;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
@@ -13,24 +12,30 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockHttpSession;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.transaction.annotation.Transactional;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.jayway.jsonpath.DocumentContext;
+import com.jayway.jsonpath.JsonPath;
 
 import shop.mtcoding.miniproject2.model.User;
 
 @Transactional
 @AutoConfigureMockMvc
 @SpringBootTest(webEnvironment = WebEnvironment.MOCK)
-public class CompanyScrapController {
+public class CompanyResumeControllerTest {
+
+    @Autowired
+    private MockMvc mvc;
 
     @Autowired
     private MockHttpSession mockSession;
 
     @Autowired
-    private MockMvc mvc;
+    private ObjectMapper om;
 
     @BeforeEach
     public void setUp() {
@@ -48,28 +53,19 @@ public class CompanyScrapController {
     }
 
     @Test
-    public void scrapInsert_test() throws Exception {
-        int resumeId = 2;
-        // String requestBody = "id=" + resumeId;
+    public void recommend_test() throws Exception {
 
-        // ResultActions resultActions = mvc
-        // .perform(post("/company/scrap/" +
-        // resumeId).content(requestBody).session(mockSession)
-        // .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE));
+        ResultActions resultActions = mvc.perform(get("/company/recommend").session(mockSession));
 
-        ResultActions resultActions = mvc
-                .perform(post("/company/scrap/" + resumeId).session(mockSession));
+        /*
+         * String data = resultActions.andReturn().getResponse().getContentAsString();
+         * String jData = om.writeValueAsString(data);
+         * DocumentContext jsonContext = JsonPath.parse(data);
+         * String rdata = jsonContext.read("$.data").toString();
+         * System.out.println("테스트: " + rdata);
+         */
+        resultActions.andExpect(jsonPath("$.msg").value("기업 인재 추천"));
+        resultActions.andExpect(jsonPath("$.data").exists());
 
-        resultActions.andExpect(status().isOk());
-    }
-
-    @Test
-    public void scrapDelete_test() throws Exception {
-        int resumeId = 1;
-
-        ResultActions resultActions = mvc
-                .perform(delete("/company/scrap/" + resumeId).session(mockSession));
-
-        resultActions.andExpect(status().isOk());
     }
 }
